+++ react/lib/JBTextarea.tsx
@@ -15,11 +15,11 @@ const JBTextarea = React.forwardRef((props:Props, ref) => {
     //we set this state so when ref change we have a render and our event listener will be updated
     const element = useRef<JBTextareaWebComponent>(null);
     useImperativeHandle(ref, () => element.current ?? undefined, [element]);
-    const {onBeforeInput,onBlur,onChange,onEnter,onFocus,onInput,onKeyDown,onKeyUp,onInit,onLoad, placeholder, name,autoHeight,disabled,error,required,validationList,value,...otherProps} = props;
-    useJBTextareaAttribute(element, {autoHeight,disabled,error,required,validationList,value});
+    const {onBeforeInput,onBlur,onChange,onEnter,onFocus,onInput,onKeyDown,onKeyUp,onInit,onLoad, placeholder, name,autoHeight,disabled,error,initialValue,required,validationList,value,...otherProps} = props;
+    useJBTextareaAttribute(element, {autoHeight,disabled,error,required,validationList});
     useEvents(element,{onBeforeInput,onBlur,onChange,onEnter,onFocus,onInput,onKeyDown,onKeyUp,onInit,onLoad});
     return (
-      <jb-textarea ref={element} name={name} label={props.label} message={props.message} placeholder={placeholder} {...otherProps}>
+      <jb-textarea ref={element} value={value ?? ""} initialValue={initialValue ?? ""} name={name} label={props.label} message={props.message} placeholder={placeholder} {...otherProps}>
         {props.children}
       </jb-textarea>
     );
@@ -31,6 +31,8 @@ type JBTextareaProps = EventProps & JBTextareaAttributes & {
     message?:string,
     name?:string,
     placeholder?:string
+    value?: string | null,
+    initialValue?: string | null,
 }
 export type Props = PropsWithChildren<JBTextareaProps> & JBElementStandardProps<JBTextareaWebComponent, keyof JBTextareaProps>;
 export {JBTextarea};